# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Shopping Analysis API",
    "description": "Backend for the Shopping Analysis project: order-history ingestion and, as of this pass, account registration/login. See `docs/design/` for the full system design; this schema is generated from the running app via `python -m app.cli export-openapi`.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/orders/upload": {
      "post": {
        "tags": [
          "orders"
        ],
        "summary": "Upload an order-history CSV",
        "description": "Parse an uploaded Walmart order-history CSV and echo its rows.\n\n:param file: The uploaded CSV file.\n:type file: UploadFile\n:raises HTTPException: If the file is missing, not a ``.csv``, not\n    valid UTF-8, or has no header row.\n:returns: The filename, column names, row count, and parsed rows.\n:rtype: dict",
        "operationId": "upload_orders_csv_orders_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_orders_csv_orders_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Upload Orders Csv Orders Upload Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Body_upload_orders_csv_orders_upload_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_orders_csv_orders_upload_post"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}